refactor: rename services/apiBuilder to services/utils
